Feeder/IM: Updating with new architecture [REFACTOR]
